Pull out role type
