+++ tensorzero-core/src/inference/types/mod.rs
@@ -61,7 +61,6 @@ use crate::serde_util::{
     deserialize_defaulted_json_string, deserialize_json_string, deserialize_optional_json_string,
 };
 use crate::tool::ToolCallInput;
-use crate::variant::chat_completion::{ASSISTANT_TEXT_TEMPLATE_VAR, USER_TEXT_TEMPLATE_VAR};
 use derive_builder::Builder;
 use extra_body::{FullExtraBodyConfig, UnfilteredInferenceExtraBody};
 use extra_headers::FullExtraHeadersConfig;
@@ -112,11 +111,13 @@ mod input_message;
 #[cfg(feature = "pyo3")]
 pub mod pyo3_helpers;
 pub mod resolved_input;
+mod role;
 pub mod storage;
 pub mod stored_input;
 pub mod streams;
 
 pub use resolved_input::ResolvedRequestMessage;
+pub use role::Role;
 pub use stored_input::{
     StoredInput, StoredInputMessage, StoredInputMessageContent, StoredRequestMessage,
 };
@@ -642,50 +643,6 @@ impl<'de> Deserialize<'de> for TextKind {
                 "Unknown key `{key}` in text content"
             ))),
         }
-    }
-}
-
-#[derive(ts_rs::TS, Clone, Copy, Debug, Deserialize, Serialize, PartialEq)]
-#[ts(export)]
-#[serde(rename_all = "snake_case")]
-#[cfg_attr(feature = "pyo3", pyclass)]
-pub enum Role {
-    User,
-    Assistant,
-}
-
-impl Role {
-    /// The template name to use for `{"type": "text", "arguments": {}}` inputs.
-    /// This will eventually be deprecated in favor of explicit `{"type": "template", "name": "user", "arguments": {}}` inputs.
-    pub fn implicit_template_name(&self) -> &'static str {
-        match self {
-            Role::User => "user",
-            Role::Assistant => "assistant",
-        }
-    }
-
-    pub fn implicit_template_var(&self) -> &'static str {
-        match self {
-            Role::User => USER_TEXT_TEMPLATE_VAR,
-            Role::Assistant => ASSISTANT_TEXT_TEMPLATE_VAR,
-        }
-    }
-}
-
-impl std::fmt::Display for Role {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Role::User => write!(f, "user"),
-            Role::Assistant => write!(f, "assistant"),
-        }
-    }
-}
-
-#[cfg(feature = "pyo3")]
-#[pymethods]
-impl Role {
-    pub fn __repr__(&self) -> String {
-        self.to_string()
     }
 }
